backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
httpx==0.25.2
pyjwt==2.8.0
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
